backend/requirements.txt
# Django y dependencias principales
Django==5.2.1
djangorestframework==3.16.0
django-filter==25.2
django-crispy-forms==2.4
crispy-bootstrap5==2024.2

# API REST y JWT
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.0

# Base de datos
psycopg[binary]>=3.2.0

# Cache y sesiones
django-redis==5.4.0
redis==5.0.7

# Exportación PDF
reportlab==4.0.9
xhtml2pdf==0.2.15

# Validaciones y utilidades
phonenumbers==8.13.46
python-dateutil==2.9.0.post0

# Desarrollo
django-debug-toolbar==4.4.6
django-extensions==3.2.3

# Testing
pytest==8.3.3
pytest-django==4.8.0
factory-boy==3.3.0

# Producción
gunicorn==23.0.0
whitenoise==6.7.0

python-dotenv>=1.0.0
dj-database-url
